Added JavetConverterConfig
